implemets icons and gives a row view to a event
